requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
